move tokenize method
